feat: add place_order, portfolio_status scripts and test infrastructure
Task 10 - place_order.py Script:
- Add CLI script for placing buy/sell orders
- Add --btc-mode for auto-discovery of BTC 15-min market
- Add --market-id and --token-id for manual override
- Add --dry-run/--live flag (defaults to dry-run for safety)
- Add pre-flight checks and confirmation prompt
- Add market order emulation via aggressive LIMIT orders

Task 11 - portfolio_status.py Script:
- Add CLI script for checking open orders and positions
- Add --market-id filter option
- Add --json output option
- Add Rich console table with order details
- Add graceful handling of read_only mode

Task 12 - Test Infrastructure:
- Add pytest.ini configuration
- Add acceptance tests in test_verification.py
- Add quick_test.sh for manual verification
- Acceptance tests verify all requirements from spec

Bug Fixes:
- Fixed order_price_min_tick_size and order_min_size types (int -> float)
- Fixed exception handling in portfolio_status
- Fixed test flag usage for --dry-run

+++ polymarket/models.py
@@ -36,8 +36,8 @@ class Market(BaseModel):
     clob_token_ids: str | None = Field(None, alias="clobTokenIds")
 
     # Trading constraints
-    order_price_min_tick_size: int | None = Field(None, alias="orderPriceMinTickSize")
-    order_min_size: int | None = Field(None, alias="orderMinSize")
+    order_price_min_tick_size: float | None = Field(None, alias="orderPriceMinTickSize")
+    order_min_size: float | None = Field(None, alias="orderMinSize")
 
     # Market data
     best_bid: float | None = Field(None, alias="bestBid")
